refactor(mitm-addon): keep buffer transport types with orchestrator
