+++ crates/runner/mitm-addon/src/usage/buffer/__init__.py
@@ -35,11 +35,8 @@ from .models import (
     ResourceFieldName,
     UsageEvent,
     UsageFlushTrigger,
-    _EnqueueWebhook,
-    _FlushOwnerLock,
-    _TimerFactory,
 )
-from .orchestrator import UsageEventBuffer
+from .orchestrator import UsageEventBuffer, _EnqueueWebhook, _FlushOwnerLock, _TimerFactory
 
 __all__ = [
     "DEFAULT_FLUSH_INTERVAL_SECONDS",
